fix the bug about axcswap

+++ src/commons/util/util.cpp
@@ -1058,10 +1058,10 @@ bool ParseInt32(const std::string& str, int32_t *out)
 
 bool  ConvertPermsToString(uint64_t perms, uint8_t total_perms_count, string& permsList) {
     ostringstream s;
-    for (int8_t i = total_perms_count - 1; i >= 0; i--) {
-        int8_t p = (perms & ( 1 << i )) > 0 ? 1 :0;
+    for (int i = total_perms_count - 1; i >= 0; i--) {
+        int p = (perms & ( 1 << i )) > 0 ? 1 :0;
         s << i << ":" << p;
-        if (i !=0 )
+        if (i != 0)
              s << ", ";
     }
